refactor: extract dashboard widgets components
